list, status, start commands, and generic & github backends
Sets up the python basic configurations and implements the list, status and
start commands.

Implements the 'generic' backend which allows for any type of backend to be
used using a pre-defined program interface.

Implements support for the Github issue system as a backend.

Initial documentation provided in form of a User Guide and basic information
in the git-tasks program itself.

issue #1 Generic backend using script interface
issue #2 Add backend for Github

Signed-off-by: Gunnar Strand <Gurra.Strand@gmail.com>